+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-4.2.0 (2014-09-04)
+4.2.0 (2014-09-06)
 ------------------
 
   * Added `stubbles\input\web\WebRequest::id()` which reads the value of an X-Request-ID header or generates a random value in case the header is missing or invalid
   * `stubbles\input\web\WebRequest::uri()` now passes the `stubbles\peer\MalformedUriException` instead of turning it into a `\RuntimeException`
+  * Implemented #66: Add warning when user agent injection is used
 
 
 4.1.1 (2014-09-01)